게시판 수정

+++ src/routes/Td.js
@@ -1,11 +1,18 @@
 /* eslint-disable */
 
 import React from "react";
+import { useNavigate } from "react-router-dom";
 
 const Td = ({ item, handleRemove, handleEdit }) => {
-  const onRemove = () => {
-    handleRemove(item.id);
-    alert("삭제 하시겠습니까?");
+  // let navigate = useNavigate();
+
+  const onRemove = (e) => {
+    if (window.confirm("삭제하시겠습니까?")) {
+      alert("삭제되었습니다.");
+      handleRemove(item.id);
+    } else {
+      alert("취소합니다.");
+    }
   };
 
   const onEdit = () => {
@@ -17,14 +24,17 @@ const Td = ({ item, handleRemove, handleEdit }) => {
       <tr className="bg-white border-2 border-gray-200">
         <td className="px-4 py-3">{item.id}</td>
         <td className="px-4 py-3">{item.username}</td>
-        <td className="px-4 py-3">{item.qa}</td>
-        {/* <td className="px-4 py-3">{item.email}</td> */}
+        <td className="px-4 py-3 qaDetail">{item.qa}</td>
         <td className="px-4 py-3">{item.phone}</td>
         <td className="text-center text-purple-400 cursor-pointer show-modal">
-          <button onClick={onEdit}>asd</button>
+          <button onClick={onEdit}>
+            <span class="material-symbols-outlined">edit</span>
+          </button>
         </td>
         <td className="text-center text-purple-400 cursor-pointer ">
-          <button onClick={onRemove}>asd</button>
+          <button onClick={onRemove}>
+            <span class="material-symbols-outlined">delete</span>
+          </button>
         </td>
       </tr>
     </>
